requirements.txt
alabaster==0.7.12
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.9
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-haystack==2.8.1
django-sphinxdoc==2.0.0
djangorestframework==3.11.0
djangorestframework-filters==1.0.0.dev0
docutils==0.16
drf-yasg==1.17.0
factory-boy==2.12.0
Faker==3.0.0
idna==2.8
imagesize==1.2.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
packaging==20.0
pdoc3==0.7.2
Pillow==7.0.0
psycopg2==2.8.4
Pygments==2.5.2
pyparsing==2.4.6
pysolr==3.8.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.7
Whoosh==2.7.4
